Split parser behavior tests
